refactor(base): split resources.rs (713 lines) into directory module
Extract test module into sibling file:

- mod.rs: resource loading logic (509 lines)
- tests.rs: unit tests (204 lines)
